pyproject.toml
[project]
name = "sauron"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "cryptography>=46.0.6",
    "fastapi>=0.116.1",
    "langchain>=1.2.13",
    "langgraph>=1.1.3",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
